move FlowAtLocation to be a dataflow abstraction
We can now use it in e.g. drop elaboration if we want to.
